Add new feature to be able to sort move & have branching factor.
There is an interesting result that appeared, where having N/B
alternating placements with a branching factor of 1, always wanting to
go closer to the origin made the algorithm get stuck.  For this reason,
I want to better understand how it got there & make the Image function
better. Perhaps I can make a .gif over time of all of the branch states.