Add a Path option for drawing the board.
This allowed me to get a knights tour, similar to (but not the same as)
the knight's tour numberphile t-shirt. The knight tries to get closer to
the origin, not the lowest value for an infinite board with a spiraling
index.. It seems not to get stuck in the first ~5k hops, unlike the
numberphile t-shirt. I kind of expected it to.

./graph.py 100 will get you a graph showing this.